+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 - Build the compiler libraries/tests in uncurried mode. https://github.com/rescript-lang/rescript-compiler/pull/6864
 - Ignore `-uncurried` command-line flag. https://github.com/rescript-lang/rescript-compiler/pull/6885
 - Cleanup: remove tracking of uncurried state in parser/printer. https://github.com/rescript-lang/rescript-compiler/pull/6888
+- Remove `%opaque` primitive. https://github.com/rescript-lang/rescript-compiler/pull/6892
 
 #### :nail_care: Polish